feat: completed jobs (very messy commit, too tired)

+++ app/src/main/java/com/realeyez/trademart/gui/components/profile/ShowcaseRow.java
@@ -36,7 +36,7 @@ public class ShowcaseRow {
         rowPanel = new LinearLayout(context);
         LinearLayout.LayoutParams row_params = new LinearLayout.LayoutParams(
                 LayoutParams.MATCH_PARENT,
-                DimensionsUtil.getScreenDensity(context) * 150);
+                (int)(DimensionsUtil.getScreenDensity(context) * 150));
         row_params.gravity = Gravity.START;
         rowPanel.setOrientation(LinearLayout.HORIZONTAL);
         rowPanel.setLayoutParams(row_params);
